Records of the bullseye puffer catch landed.

This data set provides records of the bullseye puffer catch landed on the Pacific coast of Mexico in 2023. It includes information on Biomass (live weight) for each commercial landed presentation : total weight (WT), body trunk weight (WB), and fillet weight (Wfi). The data also indicates whether the catch was Fresh or Frozen-thawed. Values are provided in Mexican pesos per landed category and per kilogram of landed weight (SIPESCA, 2024).

Usage

Botete_land

Format

## `Botete_land` A data frame with 5 obs. of 5 variables:

Category

Landed presentation.

Landed

Landed weigth by category (g).

Biomass

Converted landed category weigth to total weight (g).

Pesos

Value in Mexican pesos per landed category.

Pesos_kg

Mean value per kilogram of landed weight.

Source

<https://github.com/Macrurido/Morefi.git>

Details

Category includes five commercial landed presentations: whole fresh (WT_Fresh); body trunk fresh (WB_Fresh), body trunk Fresh-Thawed (WB_Fresh_Thawed); fillet fresh (Wfi_Fresh), and the sum of total (Total).

References

Aguirre-Villaseñor, H., Morales-Bojórquez, E., Cisneros-Mata (FISH13711).Biometric relationships as a fisheries management tool: A case study of the bullseye puffer (Sphoeroides annulatus. Tetraodontidae). Fisheries Research.

SIPESCA. 2024. Sistema de Información de Pesca y Acuacultura – SIPESCA. Comisión Nacional de Pesca y Acuacultura. <https://sipesca.conapesca.gob.mx> (accessed 7 February 2024).