fn_ARSS

The function fn_ARSS perform the Coincident Curves Test, to determine if there are significant differences between the fitted curves for each database.It is based on the Analysis of the Residual Sum of Squares (ARSS) (Chen et al. 1992).

Usage

fn_ARSS(List_TCCT, i, alfa = 0.05)

Arguments

List_TCCT

An empty list to save the matrix results from each regression.

i

An integer value indicating the ith regression analyzed.

alfa

A numerical value that defines the significance level. If this is not set, the default number is 0.05.

Value

List_TCCT calculations for ARSS are stored here

Details

The residual sum of squares (RSS) and the degrees of freedom (DF) for each fitted regression are stored previously in the List_TCCT list. Calculations for ARSS are also stored here.

Inside the function, the RSS and DF for the joined sample are calculated to perform the F test for two tails alpha/2. The decision criteria is performed: "*" if the p-value is less than or equal to alpha or "NS" if the p-value is greater than alpha.

The function requires defining.

See also

Examples

if (FALSE) { # \dontrun{
i <- 1
List_TCCT <- vector("list", i)

ARSS <- fn_ARSS(List_TCCT, i, alfa = 0.05)
} # }